refactor: move base components into `base` directory
