+++ src/components/auth/SignUpForm.tsx
@@ -5,8 +5,8 @@ import Terms from 'constants/Terms';
 import { ReactComponent as EditIcon } from 'assets/auth/mingcute_pencil-fill.svg';
 import NicknameInput from 'components/auth/NicknameInput';
 import ProfileSelectModal from 'components/auth/ProfileSelectModal';
-import Button from 'components/Button';
-import Typography from 'components/Typography';
+import Button from 'components/base/Button';
+import Typography from 'components/base/Typography';
 import useSignUp from 'hooks/api/useSignUp';
 import useAccessToken from 'hooks/useAccessToken';
 import { FormProvider, useForm } from 'react-hook-form';
