+++ src/pages/LoginPage.tsx
@@ -3,7 +3,7 @@ import Terms from 'constants/Terms';
 import { ReactComponent as PoomLogo } from 'assets/app/poom.svg';
 import KakaoLoginImage from 'assets/auth/kakaoLoginMedium.png';
 import { ReactComponent as BackIcon } from 'assets/components/navigate/Back.svg';
-import Typography from 'components/Typography';
+import Typography from 'components/base/Typography';
 import useQueryKakaoLogin from 'hooks/useQueryKakaoLogin';
 import useRedirectPath from 'hooks/useRedirectPath';
 import { useTranslation } from 'react-i18next';
