+++ src/pages/error/NotFoundErrorPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import PathName from 'constants/PathName';
-import Typography from 'components/Typography';
+import Typography from 'components/base/Typography';
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router';
